+++ app/Http/Controllers/FeatureController.php
@@ -4,19 +4,26 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Feature;
+use App\Models\Room;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class FeatureController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Room::class) && ($request->user()->cannot('update', Room::class) || $request->user()->cannot('create', Room::class))) {
+            return abort(403);
+        }
         $features = Feature::with('category')->get();
-        return Inertia::render('Admin/Rooms/Features', ['features' => fn () => $features, 'categorys' => fn () => Category::all()]);
+        return Inertia::render('Admin/Rooms/Features', ['features' => fn() => $features, 'categorys' => fn() => Category::all()]);
     }
 
     public function store(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Room::class) && ($request->user()->cannot('update', Room::class) || $request->user()->cannot('create', Room::class))) {
+            return abort(403);
+        }
         request()->validate(
             [
                 'features_name' => 'required|string',
@@ -34,6 +41,9 @@ class FeatureController extends Controller
 
     public function update(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Room::class) && ($request->user()->cannot('update', Room::class) || $request->user()->cannot('create', Room::class))) {
+            return abort(403);
+        }
         request()->validate(
             [
                 'features_name' => 'required|string',
@@ -50,8 +60,11 @@ class FeatureController extends Controller
         return redirect(route('features.index'))->with('message', ['status' => 'success', 'message' => 'Caractéristique modifier avec succès']);
     }
 
-    public function destroy(string $id)
+    public function destroy(string $id, Request $request)
     {
+        if ($request->user()->cannot('viewAny', Room::class) && ($request->user()->cannot('update', Room::class) || $request->user()->cannot('create', Room::class))) {
+            return abort(403);
+        }
         Feature::where('feature_id', $id)->delete();
         return redirect()->back()->with('message', ['status' => 'success', 'message' => 'Caractéristique supprimier avec succès']);
     }
